refactor: standardise on named exports and drop the barrel files
The fourteen component barrels had grown five different shapes between
them: `export *` plus `export { default }`, explicit named exports, a bare
`export { default }`, and one that imported a default only to re-export it
on the next line.

Beyond the inconsistency, `export *` does not declare a public API. It
re-exports whatever the module happens to export now and in future, so a
new export joins the public surface without anyone deciding to put it
there, and bundlers have less to work with when tree-shaking. Pairing it
with `export { default }` compounds the confusion, since the second line
exists only to compensate for what the first deliberately omits.

Remove the barrels entirely and import each component from its own file.
For single-file components a barrel buys nothing but indirection and the
risk of a cycle. Convert every default export to a named one along the
way, App included, so renaming a component is a rename everywhere rather
than a silent rebinding at each import site.

The default exports in vite-env.d.ts stay as they are - that is how Vite
types asset imports. src/types/index.ts is a real module, not a barrel,
and is untouched.
