--- src/components/AboutMe/index.ts
@@ -1,2 +1,0 @@
-export * from './AboutMe';
-export { default } from './AboutMe';
