--- src/components/Card/index.ts
@@ -1,15 +1,0 @@
-export {
-  Card,
-  CardContent,
-  CardDecoration,
-  CardFooter,
-  CardHeader,
-} from './Card';
-
-export type {
-  CardContentProps,
-  CardDecorationProps,
-  CardFooterProps,
-  CardHeaderProps,
-  CardProps,
-} from './Card';
