+++ src/components/ContactInfo/ContactInfo.test.tsx
@@ -1,4 +1,4 @@
-import ContactInfo from '@components/ContactInfo';
+import { ContactInfo } from '@components/ContactInfo/ContactInfo';
 import { describe, expect, it } from '@jest/globals';
 import { render, screen } from '@testing-library/react';
 import { axe } from 'jest-axe';
