--- src/components/ContactInfo/index.ts
@@ -1,2 +1,0 @@
-export * from './ContactInfo';
-export { default } from './ContactInfo';
