--- src/components/Experience/index.ts
@@ -1,2 +1,0 @@
-export * from './Experience';
-export { default } from './Experience';
