--- src/components/FlipCard/index.ts
@@ -1,2 +1,0 @@
-export { FlipCard } from './FlipCard';
-export type { FlipCardProps } from './FlipCard';
