--- src/components/Footer/index.ts
@@ -1,2 +1,0 @@
-export * from './Footer';
-export { default } from './Footer';
