+++ src/components/FunFacts/FunFacts.test.tsx
@@ -1,4 +1,4 @@
-import FunFacts from '@components/FunFacts';
+import { FunFacts } from '@components/FunFacts/FunFacts';
 import { describe, expect, it } from '@jest/globals';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
