--- src/components/Header/index.ts
@@ -1,2 +1,0 @@
-export * from './Header';
-export { default } from './Header';
