+++ src/components/LearningShelf/LearningShelf.test.tsx
@@ -5,7 +5,7 @@ import { axe } from 'jest-axe';
 
 import { data } from '@/data';
 
-import LearningShelf from './LearningShelf';
+import { LearningShelf } from './LearningShelf';
 
 describe('LearningShelf', () => {
   it('renders with correct heading', () => {
