+++ src/components/LearningShelf/LearningShelf.tsx
@@ -6,7 +6,7 @@ import type { LearningItem } from '@/types';
 
 import { LearningShelfItem } from './LearningShelfItem';
 
-function LearningShelf(): React.JSX.Element {
+export function LearningShelf(): React.JSX.Element {
   return (
     <section
       className="w-full flex-1 flex flex-col gap-6 px-4 py-12 items-center text-text-primary dark:text-text-primary-dark"
@@ -39,5 +39,3 @@ function LearningShelf(): React.JSX.Element {
     </section>
   );
 }
-
-export default LearningShelf;
