--- src/components/LearningShelf/index.ts
@@ -1,3 +1,0 @@
-import LearningShelf from './LearningShelf';
-
-export default LearningShelf;
