--- src/components/Projects/index.ts
@@ -1,2 +1,0 @@
-export * from './Projects';
-export { default } from './Projects';
