--- src/components/TagList/index.ts
@@ -1,2 +1,0 @@
-export * from './TagList';
-export { default } from './TagList';
