+++ src/components/ThemeToggle/ThemeToggle.test.tsx
@@ -1,4 +1,4 @@
-import ThemeToggle from '@components/ThemeToggle';
+import { ThemeToggle } from '@components/ThemeToggle/ThemeToggle';
 import { describe, expect, it } from '@jest/globals';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
