+++ src/main.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import '@styles/index.css';
 import ReactDOM from 'react-dom/client';
 
-import App from '@/App';
+import { App } from '@/App';
 import {
   applyTheme,
   getInitialTheme,
